[API] В роутах albums и tracks были не большие поправки. [FRONT] Был компоненте UserMenu были не большие изменения. Далее в диретории features была создана субдиректория admin с контейнером AdminContainer и в файле App.tsx был прописан для него роут.  Далее были не большие изменения и проверки в сущностях artists, albums, tracks и trackHistories.

+++ frontend/src/features/artists/component/ArtistsCards/ArtistsCard/ArtistsCard.tsx
@@ -11,8 +11,9 @@ import { useNavigate } from 'react-router-dom';
 import Button from '@mui/material/Button';
 import { useAppDispatch, useAppSelector } from '../../../../../app/hooks.ts';
 import { userFromSlice } from '../../../../users/usersSlice.ts';
-import { deleteArtist, getArtists } from '../../../artistsThunk.ts';
+import { deleteArtist, getArtists, publishArtist } from '../../../artistsThunk.ts';
 import { toast } from 'react-toastify';
+import Box from '@mui/material/Box';
 
 interface Props {
   artist: IArtist;
@@ -29,11 +30,29 @@ const ArtistsCard:React.FC<Props> = ({artist}) => {
   }
 
   const deleteThisArtist = async (artistId: string) => {
-    if (user) {
+    if (user && user.role === 'user') {
       await dispatch(deleteArtist({artistId, token: user.token}));
       await dispatch(getArtists());
       toast.success('Artist was successfully deleted!');
       navigate('/');
+    }
+  };
+
+  const deleteAdminTheArtist = async (artistId: string) => {
+    if (user && user.role === 'admin') {
+      await dispatch(deleteArtist({artistId, token: user.token}));
+      await dispatch(getArtists());
+      toast.success('Artist was successfully deleted!');
+      navigate('/admin');
+    }
+  };
+
+  const publishAdminTheArtist = async (artistId: string) => {
+    if (user && user.role === 'admin') {
+      await dispatch(publishArtist({artistId, token: user.token})).unwrap();
+      await dispatch(getArtists());
+      toast.success('Artist was successfully published!');
+      navigate('/admin');
     }
   };
 
@@ -55,8 +74,18 @@ const ArtistsCard:React.FC<Props> = ({artist}) => {
           {artist.name}
         </Typography>
         {artist.description ? <Typography level="body-sm">{artist.description}</Typography> : null}
-        {!artist.isPublished && (
-          <Button variant="text" sx={{width: '100px'}} onClick={() => deleteThisArtist(artist._id)}>Delete</Button>)}
+        {user && user.role !== 'admin' && !artist.isPublished ? (
+          <Button variant="text" sx={{width: '100px'}} onClick={() => deleteThisArtist(artist._id)}>Delete</Button>)
+        :
+          <>
+            {user !== null && user.role === 'admin' && (
+            <Box>
+              <Button variant="text" sx={{width: '100px'}} onClick={() => deleteAdminTheArtist(artist._id)}>Delete</Button>
+              <Button variant="text" sx={{width: '100px'}} onClick={() => publishAdminTheArtist(artist._id)}>Publish</Button>
+            </Box>
+          )}
+          </>
+        }
       </CardContent>
       {!artist.isPublished && (
         <CardOverflow
